Tests/LibWeb/Text/expected/hit_testing/overflow-scroll.txt

<P id="line-6" >
<SPAN id="line-8" >
